splitter Beregning komponent i en komponent for hver fane
