+++ src/mocks/data/unleash-enable-detaljert-fane.json
@@ -1,3 +1,3 @@
 {
-  "enabled": false
+  "enabled": true
 }
