+++ src/pages/Beregning/Beregning.module.scss
@@ -33,21 +33,6 @@
   left: 0;
   right: 0;
 
-  &__hasMargin {
-    margin-top: var(--a-spacing-4);
-  }
-
-  &__hasMinheight {
-    @media (min-width: $a-breakpoint-md) {
-      min-height: 33vh;
-    }
-  }
-
-  &__white {
-    background-color: var(--a-white);
-    transition: background-color 0.15s ease;
-  }
-
   &__lightblue {
     background-color: var(--a-deepblue-50);
   }
