+++ src/pages/Beregning/BeregningAvansert.tsx
@@ -1,0 +1,5 @@
+import React from 'react'
+
+export const BeregningAvansert: React.FC = () => {
+  return <p>avansert visning</p>
+}
