+++ src/pages/Beregning/BeregningEnkel.module.scss.d.ts
@@ -1,10 +1,10 @@
 import globalClassNames from '../../style.d'
 declare const classNames: typeof globalClassNames & {
-  readonly beregning: 'beregning'
   readonly container: 'container'
   readonly container__hasPadding: 'container__hasPadding'
-  readonly toggle: 'toggle'
   readonly background: 'background'
-  readonly background__lightblue: 'background__lightblue'
+  readonly background__hasMargin: 'background__hasMargin'
+  readonly background__hasMinheight: 'background__hasMinheight'
+  readonly background__white: 'background__white'
 }
 export = classNames
