+++ src/pages/Beregning/index.ts
@@ -1,1 +1,2 @@
 export { Beregning } from './Beregning'
+export { BeregningAvansert } from './BeregningAvansert'
